webDev/sorteio_app/requirements.txt
flask
gunicorn
psycopg2-binary
supabase

